This patch contains two items:
The first is to disable strict SSL checking. I was getting constant failures attempting to connect
to my capsule domain. Feel free to disregard although it does seem like I shouldn't be the only
person encountering this.

The second is to enable the returning of an id from a request response. This is done by using a
match against the location returned in the headers section of the HTTP_Request2 response object.